+++ packages/ui-helpers/src/Footer.tsx
@@ -11,8 +11,10 @@ import {
   DiscordIcon,
   CommonWealthIcon,
   CodeIcon,
+  MediumIcon,
 } from "icons";
 import { PRIVACY_POLICY_URL } from "constants-helper";
+import { MEDIUM_URL } from "constants-helper/src/constants";
 
 export const Footer = ({
   handleCookies,
@@ -95,6 +97,14 @@ export const Footer = ({
             aria-label="discord telegram"
           >
             <TelegramIcon />
+          </Link>
+          <Link
+            target="_blank"
+            rel="noreferrer"
+            href={MEDIUM_URL}
+            aria-label="medium evmos"
+          >
+            <MediumIcon width={24} height={24} />
           </Link>
           <Link
             target="_blank"