+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- (chore) fse-799 | packages/copilot 1.0.10 | Use wagmi value instead of redux value
 - (fix) fse-816 | packages/ui-helpers 1.0.17 packages/icons 1.0.10 | Replace Twitter logo to X logo in the footer
 - (fix) fse-831 | apps/assets 1.0.39 | Fix usdc.grv image
 - (fix) fse-837 | apps/vesting 1.1.4 | Fix error in account details page for vesting